--- frontend/src/commons/color.ts
@@ -1,18 +1,0 @@
-interface IColor {
-  [key: string]: string;
-}
-
-export const chkColors: IColor = {
-  orange: "#edb600",
-  blue: "#66C9DB",
-  skyblue: "#A8E1E3",
-  black: "#565A62",
-  gray: "#7f8287",
-  shadow: "#14141433",
-  white: "#ffffff",
-  placeholder: "#b0bec5",
-  aqua: "#b3dfe2",
-  darkenAqua: "#50AAB7",
-  error: "#ff4b42",
-  disabled: "#cccccc"
-};
