+++ frontend/src/components/atoms/HeroArea/hero.jpg
