+++ frontend/src/components/atoms/TweetStream.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import styled from "styled-components";
 import { Timeline } from "react-twitter-widgets";
 
-import { title } from "./../commons/style";
+import { title } from "../../commons/style";
 import { color } from "@/constants/styles";
 
 export default () => (
